va: long -> int32

+++ src/cmd/va/a.h
@@ -26,11 +26,13 @@ typedef	struct	Hist	Hist;
 #define	STRINGSZ	200
 #define	NMACRO		10
 
+// kengo: use of int32 instead of long
+
 struct	Sym
 {
 	Sym*	link;
 	char*	macro;
-	long	value;
+	int32	value;
 	ushort	type;
 	char	*name;
 	char	sym;
@@ -62,7 +64,7 @@ EXTERN	struct
 struct	Gen
 {
 	Sym*	sym;
-	long	offset;
+	int32	offset;
 	short	type;
 	short	reg;
 	short	name;
@@ -74,8 +76,8 @@ struct	Hist
 {
 	Hist*	link;
 	char*	name;
-	long	line;
-	long	offset;
+	int32	line;
+	int32	offset;
 };
 #define	H	((Hist*)0)
 
@@ -100,9 +102,9 @@ EXTERN	char**	include;
 EXTERN	Io*	iofree;
 EXTERN	Io*	ionext;
 EXTERN	Io*	iostack;
-EXTERN	long	lineno;
+EXTERN	int32	lineno;
 EXTERN	int	nerrors;
-EXTERN	long	nhunk;
+EXTERN	int32	nhunk;
 EXTERN	int	nosched;
 EXTERN	int	ninclude;
 
@@ -113,17 +115,16 @@ EXTERN	Gen	nullgen;
 EXTERN	char*	outfile;
 EXTERN	int	pass;
 EXTERN	char*	pathname;
-EXTERN	long	pc;
+EXTERN	int32	pc;
 EXTERN	int	peekc;
 EXTERN	int	sym;
 //kengo: was char	symb[NSYMB]; now dynamically alloc'ed in lexbody
 EXTERN	char*	symb;
 EXTERN	int	thechar;
 EXTERN	char*	thestring;
-EXTERN	long	thunk;
+EXTERN	int32	thunk;
 EXTERN	Biobuf	obuf;
 
-// kengo: use of int32 instead of long in a few protos below
 void*	alloc(int32);
 void*	allocn(void*, int32, int32);
